Add spring security and lombok

+++ src/main/java/com/igormascarenhas/amivulnerable/device/DeviceController.java
@@ -39,7 +39,7 @@ public class DeviceController {
     }
 
     @DeleteMapping(path = "/device/{deviceId}")
-    public void deleteDevice(@PathVariable("deviceId") Integer id) {
+    public void deleteDevice(@PathVariable("deviceId") Long id) {
         deviceService.deleteDevice(id);
     }
 }
